px2ce: require が使えない謎の現象を解決。
