+++ app/index.html
@@ -49,8 +49,8 @@
 		<script src="./common/broccoli-field-table/dist/broccoli-field-table.js"></script>
 
 		<!-- Pickles 2 Contents Editor -->
-		<link rel="stylesheet" href="./common/pickles2-contents-editor/dist/pickles2-contents-editor.css" />
-		<script src="./common/pickles2-contents-editor/dist/pickles2-contents-editor.js"></script>
+		<link rel="stylesheet" href="./common/pickles2-contents-editor/dist/pickles2-contents-editor.css" charset="utf-8" />
+		<script src="./common/pickles2-contents-editor/dist/pickles2-contents-editor.js" charset="utf-8"></script>
 
 		<script id="template-selectProject-page" type="text/template">
 			<div class="cont_top_jumbotron">
