+++ app/mods/editor/editor_px2ce.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta charset="UTF-8" />
+		<title>Editor - Pages - Pickles 2 Desktop Tool</title>
+
+		<meta name="viewport" content="width=device-width" />
+		<meta name="keywords" content="" />
+		<meta name="description" content="" />
+
+		<!-- jQuery -->
+		<script src="../../common/scripts/jquery-1.11.1.min.js" type="text/javascript"></script>
+
+		<!-- Ace Editor -->
+		<script src="../../common/ace-builds/src-noconflict/ace.js" type="text/javascript"></script>
+
+		<!-- underscore -->
+		<script src="../../common/underscore/underscore-min.js" type="text/javascript"></script>
+
+		<!-- Bootstrap -->
+		<link rel="stylesheet" href="../../common/bootstrap/css/bootstrap.min.css">
+		<script src="../../common/bootstrap/js/bootstrap.min.js"></script>
+
+		<!-- Keypress -->
+		<script src="../../common/dmauro-Keypress/keypress.js"></script>
+
+		<!-- normalize & FESS -->
+		<link rel="stylesheet" href="../../common/styles/contents.css" type="text/css" />
+
+		<!-- broccoli-html-editor -->
+		<link rel="stylesheet" href="../../common/broccoli-html-editor/client/dist/broccoli.css" />
+		<script src="../../common/broccoli-html-editor/client/dist/broccoli.js"></script>
+
+		<!-- broccoli-field-table -->
+		<script src="../../common/broccoli-field-table/dist/broccoli-field-table.js"></script>
+
+		<!-- broccoli-field-href -->
+		<script src="../../common/broccoli/broccoli-field-href/client.js"></script>
+
+		<!-- Pickles 2 Contents Editor -->
+		<link rel="stylesheet" href="../../common/pickles2-contents-editor/dist/pickles2-contents-editor.css" />
+		<script src="../../common/pickles2-contents-editor/dist/pickles2-contents-editor.js"></script>
+
+		<!-- Local Resources -->
+		<link rel="stylesheet" href="./editor_px2ce_files/style.css" type="text/css" />
+		<script src="./editor_px2ce_files/script.js" charset="utf-8"></script>
+		<script src="./editor_px2ce_files/server.js" charset="utf-8"></script>
+
+	</head>
+	<body>
+		<div id="canvas"></div>
+	</body>
+</html>
