+++ app/mods/editor/editor_px2ce_files/style.css
@@ -1,0 +1,7 @@
+@charset "utf-8";
+html, body{
+	height:100%;
+}
+#canvas{
+	height:100%;
+}
