+++ app/mods/editor/index_files/style.css
@@ -1,7 +1,1 @@
 @charset "utf-8";
-html, body{
-	height:100%;
-}
-#canvas{
-	height:100%;
-}
